Fix for Screen Rotate resetting the search results.

+++ app/src/main/java/com/andrewchelladurai/simplebible/FragmentSearch.java
@@ -70,24 +70,12 @@ public class FragmentSearch
         mButton.setOnClickListener(this);
 
         RecyclerView recyclerView = (RecyclerView) view.findViewById(R.id.frag_search_results);
-
-        // FIXME: 9/7/16 Rotate resets everything.
-        ListSearch.populate(null);
+        if (savedInstanceState == null) {
+            ListSearch.truncate();
+        }
         mListAdapter = new AdapterSearchList(ListSearch.getEntries(), this);
-        resetButtonClicked();
         recyclerView.setAdapter(mListAdapter);
-
         return view;
-    }
-
-    private void resetButtonClicked() {
-        ListSearch.truncate();
-        mInput.setText("");
-        mInput.setError(null);
-        mLabel.setError(null);
-        mButton.setText(getString(R.string.button_search_text));
-        mListAdapter.notifyDataSetChanged();
-        mInput.requestFocus();
     }
 
     @Override public void onClick(View v) {
@@ -125,6 +113,16 @@ public class FragmentSearch
         ListSearch.populate(list);
         mListAdapter.notifyDataSetChanged();
         mButton.setText(getString(R.string.button_search_reset));
+    }
+
+    private void resetButtonClicked() {
+        ListSearch.truncate();
+        mInput.setText("");
+        mInput.setError(null);
+        mLabel.setError(null);
+        mButton.setText(getString(R.string.button_search_text));
+        mListAdapter.notifyDataSetChanged();
+        mInput.requestFocus();
     }
 
     public void buttonSaveClicked(ListSearch.Entry entry) {
